Module Topkg_care

Topkg package care.

Tools to help the package developer in the life cycle of the package. Most of these tools can be invoked directly from the command line via the topkg binary installed by the topkg-care package.

WARNING. Do not use this API in your package description file, use only Topkg. This API was not thoroughly designed, is not stable and may change even between minor versions of topkg. Use at your own risk.

v1.0.4 - homepage

Helpers

module Text : sig ... end

Text processing helpers.

module Pp : sig ... end

Pretty printers.

module Opam : sig ... end

opam helpers.

module OCamlbuild : sig ... end

ocamlbuild helpers.

module OCamlfind : sig ... end

ocamlfind helpers.

module Archive : sig ... end

Archive file creation.

Package care

module Pkg : sig ... end

Package description.

module Delegate : sig ... end

Package delegate.