Module Topkg_care.Delegate

Package delegate.

Publish

val publish_distrib : Pkg.t -> msg:string -> archive:Fpath.t -> (unit, Bos_setup.R.msg) Bos_setup.result

publish_distrib p ~msg ~archive publishes the distribution archive archive of package p with publication message msg.

val publish_doc : Pkg.t -> msg:string -> docdir:Fpath.t -> (unit, Bos_setup.R.msg) Bos_setup.result

publish_distrib p ~msg ~docdir publishes the documentation directory docdir of package p with publication message msg.

val publish_alt : Pkg.t -> kind:string -> msg:string -> archive:Fpath.t -> (unit, Bos_setup.R.msg) Bos_setup.result

publish_alt p ~kind ~msg ~archive publishes the artefact kind for distribution archive archive of package p with publication message msg.

Helpers

val publish_in_git_branch : remote:string -> branch:string -> name:string -> version:string -> docdir:Fpath.t -> dir:Fpath.t -> (unit, Bos_setup.R.msg) Bos_setup.result

publish_in_git_branch ~remote ~branch ~name ~version ~docdir ~dir publishes the documentation directory docdir of a package named name at version version by replacing the dir sub-directory of the branch branch of the current working directory git repository (use "." to copy the docdir at the root directory of the branch) and pushes the branch to remote.

Note. The publication procedure first checkouts the gh-pages in a temporary clone located in the Fpath.parent directory of docdir. The branch branch of this clone is then pushed to the current working git repository, whose branch branch is then pushed to the remote repository.

Issues

val issue_list : Pkg.t -> (unit, Bos_setup.R.msg) Bos_setup.result

issue_list p outputs the issue list on stdout.

val issue_show : Pkg.t -> id:string -> (unit, Bos_setup.R.msg) Bos_setup.result

issue_show p ~id outputs information about issue id on stdout.

val issue_open : Pkg.t -> title:string -> body:string -> (unit, Bos_setup.R.msg) Bos_setup.result

issue_open p ~title ~body create a new issue with title title and description body body.

val issue_close : Pkg.t -> id:string -> msg:string -> (unit, Bos_setup.R.msg) Bos_setup.result

issue_close p ~id ~msg closes issue id with message msg.