Module Topkg_care.OCamlbuild

ocamlbuild helpers.

Command

val cmd : Bos_setup.Cmd.t

cmd is a command for ocamlbuild looked up using Topkg.Conf.tool "ocamlbuild" `Host_os.

Packages

val package_tags : ?roots:bool -> Fpath.t -> (Bos_setup.String.set, Bos_setup.R.msg) Bos_setup.result

packages ~roots f is the set of packages identifiers mentioned by the package tags of the _tags file f. If roots is true (defaults to false) only root packages, i.e. the identifier before the first '.', are in the set.

Warning. This is a very dumb parsing that simply looks up for all "package($ID)" and "package($ID0[, ]*$ID1...)" patterns in the _tags file.