Module Topkg_care.Opam

opam helpers.

Command

val cmd : Bos_setup.Cmd.t

cmd is a command for opam looked up using Topkg.Conf.tool "opam" `Host_os.

Publish

val ensure_publish : unit -> (unit, Bos_setup.R.msg) Bos_setup.result

ensure_publish () makes sure opam-publish is in the executable search PATH.

val submit : ?msg:string -> opam_file:Fpath.t -> (unit, Bos_setup.R.msg) Bos_setup.result

submit ~opam_file submits the opam file opam_file with opam-publish and submission message msg (if any) to the OCaml opam repository.

Packages

val ocaml_base_packages : Bos_setup.String.set

ocaml_base_packages are the base opam packages distributed with OCaml: "base-bigarray", "base-bytes", "base-threads", "base-unix".

Files

module File : sig ... end

opam files

module Descr : sig ... end

descr files.

module Url : sig ... end

url files.