Module Opam.Descr

descr files.

Descr file

type t = string * string

The type for opam descr files, the package synopsis and the description.

val of_string : string -> (t, Bos_setup.R.msg) Bos_setup.result

of_string s is a description from the string s.

val to_string : t -> string

to_string d is d as a string.

val to_opam_fields : t -> string

to_opam_fields d is d as synopsis and description opam v2 fields.

val of_readme : ?flavour:Text.flavour -> string -> (t, Bos_setup.R.msg) Bos_setup.result

of_readme r extracts an opam description file from a readme r with a certain structure.

val of_readme_file : Fpath.t -> (t, Bos_setup.R.msg) Bos_setup.result

of_readme_file f extracts an opam description file from a readme file f using Text.flavour_of_fpath and of_readme.