Module Opam.File

opam files

opam file

val field_names : Bos_setup.String.set

field_names is the maximal domain of the map returned by fields, excluding extension fields (not yet supported by opam-lib 1.2.2).

val fields : Fpath.t -> (string list Bos_setup.String.map, Bos_setup.R.msg) Bos_setup.result

fields f returns a simplified model of the fields of the opam file f. The domain of the result is included in field_names. Note that the depends: and depopts: fields are returned without version constraints.

Dependencies

val deps : ?opts:bool -> string list Bos_setup.String.map -> Bos_setup.String.set

deps ~opts fields returns the packages mentioned in the depends: fields, if opts is true (default) those from depopts: are added aswell.